requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-socketio[asyncio]==5.11.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
redis[hiredis]==5.0.4
celery==5.3.6
python-multipart==0.0.9
aiofiles==23.2.1
Pillow==10.3.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
greenlet==3.0.3
slowapi==0.1.9
email-validator==2.1.1
